move dedup iterators to dedup package
